+++ src/share.tsx
@@ -2,14 +2,15 @@ import { ChangeEvent, useEffect, useState } from 'react'
 import { SpaceDID, useW3 } from '@w3ui/react'
 import { extract } from '@ucanto/core/delegation'
 import type { PropsWithChildren } from 'react'
-import type { Delegation } from '@ucanto/interface'
+import type { Capabilities, Delegation } from '@ucanto/interface'
 import { SpacePreview } from './components/SpaceCreator'
-import { H2 } from '@/components/Text'
+import { H2, H3 } from '@/components/Text'
 import CopyButton from './components/CopyButton'
 import Tooltip from './components/Tooltip'
 import { ArrowDownOnSquareStackIcon, CloudArrowDownIcon, PaperAirplaneIcon, InformationCircleIcon } from '@heroicons/react/24/outline'
 import * as DIDMailTo from '@web3-storage/did-mailto'
 import { DID } from '@ucanto/core'
+import { logAndCaptureError } from './sentry'
 
 function Header(props: PropsWithChildren): JSX.Element {
   return (
@@ -64,64 +65,16 @@ export function ShareSpace({ spaceDID }: { spaceDID: SpaceDID }): JSX.Element {
       throw new Error(`Client not found`)
     }
 
-    const currentSpace = client.agent.currentSpace()
-    try {
-      const space = client.spaces().find(s => s.did() === spaceDID)
-      if (!space) {
-        throw new Error(`Could not find space to share`)
-      }
-
-      const delegatedEmail = DIDMailTo.email(email)
-
-      // FIXME (fforbeck): enable shareSpace function call after @w3ui/react lib is updated to v2.4.0 and the issue with blobs are solved
-      // const delegation = await client.shareSpace(delegatedEmail, space.did())
-
-      // Make sure the agent is using the shared space before delegating
-      await client.agent.setCurrentSpace(spaceDID)
-
-      // Delegate capabilities to the delegate account to access the **current space**
-      const delegation = await client.createDelegation(
-        {
-          did: () => DIDMailTo.fromEmail(delegatedEmail),
-        },
-        [
-          'space/*',
-          'store/*',
-          'upload/*',
-          'access/*',
-          'usage/*',
-          // @ts-expect-error (FIXME: https://github.com/storacha/w3up/issues/1554)
-          'filecoin/*',
-        ], {
-          expiration: Infinity
-        }
-      )
-
-      const sharingResult = await client.capability.access.delegate({
-        space: spaceDID,
-        delegations: [delegation],
-      })
-
-      if (sharingResult.error) {
-        throw new Error(
-          `failed to share space with ${delegatedEmail}: ${sharingResult.error.message}`,
-          {
-            cause: sharingResult.error,
-          }
-        )
-      }
-
-      const next = { email: delegatedEmail, capabilities: delegation.capabilities.map(c => c.can) }
-      updateSharedEmails([next])
-      setValue('')
-    } catch (err) {
-      console.error(err)
-    } finally {
-      // Reset to the original space if it was different
-      if (currentSpace && currentSpace !== spaceDID) {
-        await client.agent.setCurrentSpace(currentSpace)
-      }
+    const space = client.spaces().find(s => s.did() === spaceDID)
+    if (!space) {
+      throw new Error(`Could not find space to share`)
     }
+
+    const delegatedEmail = DIDMailTo.email(email)
+    const delegation: Delegation<Capabilities> = await client.shareSpace(delegatedEmail, space.did())
+    const next = { email: delegatedEmail, capabilities: delegation.capabilities.map(c => c.can) }
+    updateSharedEmails([next])
+    setValue('')
   }
 
   async function makeDownloadLink(did: string): Promise<string> {
@@ -235,12 +188,16 @@ export function ShareSpace({ spaceDID }: { spaceDID: SpaceDID }): JSX.Element {
             Shared With:
           </p>
           <ul>
-            {sharedEmails.map(({ email, capabilities }) => (
+            {sharedEmails.map(({ email, capabilities }, i) => (
               <li key={email} className="flex flex-col sm:flex-row items-start sm:items-center space-y-2 sm:space-y-0 sm:space-x-2 w-full mt-1">
                 <span className="flex items-center w-full">
                   <span className="truncate mt-1">{email}</span>
-                  <Tooltip title="Capabilities" text={capabilities}>
-                    <InformationCircleIcon className='h-5 w-5 ml-1' />
+                  <InformationCircleIcon className={`h-5 w-5 ml-1 share-capabilities-${i}`} />
+                  <Tooltip anchorSelect={`.share-capabilities-${i}`}>
+                    <H3>Capabilities</H3>
+                    {capabilities.map((c, j) => (
+                      <p key={j}>{c}</p>
+                    ))}
                   </Tooltip>
                 </span>
               </li>
@@ -267,14 +224,14 @@ export function ImportSpace() {
       }
       delegation = res.ok
     } catch (err) {
-      console.error(err)
+      logAndCaptureError(err)
       return
     }
     try {
       await client.addSpace(delegation)
       setProof(delegation)
     } catch (err) {
-      console.error(err)
+      logAndCaptureError(err)
     }
   }